workshop-main/requirements.txt
fastapi==0.116.1
langchain_google_genai==2.1.8
pandas==2.3.1
python-dotenv==1.1.1
uvicorn==0.35.0
langchain==0.3.27